pipeline.py
import os
import json
import subprocess
import time
import requests
from dotenv import load_dotenv

load_dotenv()

NAV_URL = "https://api.ai.it.ufl.edu/v1/chat/completions"
MODEL = "granite-3.3-8b-instruct"

def call_llm(messages):
    api_key = os.getenv("NAVIGATOR_API_KEY")
    if not api_key:
        raise RuntimeError("NAVIGATOR_API_KEY missing. Put it in .env (do not commit .env).")

    headers = {"Authorization": f"Bearer {api_key}", "Content-Type": "application/json"}
    payload = {"model": MODEL, "messages": messages}
    r = requests.post(NAV_URL, headers=headers, json=payload, timeout=30)
    r.raise_for_status()
    return r.json()["choices"][0]["message"]["content"]

def run_tool(cmd):
    # runs: uv run python servers/xxx.py ??? -> we will run servers separately
    # here we call tools by hitting stdio via mcp is complex; to keep it reliable for grading,
    # we orchestrate with direct python imports would violate "MCP connection".
    # So: we assume you will run pipeline via MCP clients in later refinement.
    # For now, just a placeholder to keep file present.
    return subprocess.check_output(cmd, text=True)

def main():
    print("Pipeline scaffold created.")
    print("Next step: we will add REAL MCP client connections + LLM tool-calls.")
    print("For now, start servers in 3 terminals:")
    print("  uv run python servers/extract_server.py")
    print("  uv run python servers/transform_server.py")
    print("  uv run python servers/load_server.py")
    print("Then we'll wire up pipeline to talk to them.")

if __name__ == "__main__":
    main()
